requirements.txt
alembic==1.7.5
amqp==5.0.6
anyio==3.4.0
apache-airflow==2.2.0
apache-airflow-providers-amazon==2.2.0
apache-airflow-providers-celery==2.1.0
apache-airflow-providers-cncf-kubernetes==2.0.3
apache-airflow-providers-ftp==2.0.1
apache-airflow-providers-http==2.0.1
apache-airflow-providers-imap==2.0.1
apache-airflow-providers-postgres==2.3.0
apache-airflow-providers-redis==2.0.1
apache-airflow-providers-slack==4.1.0
apache-airflow-providers-sqlite==2.0.1
apispec==3.3.2
argcomplete==1.12.3
attrs==20.3.0
Babel==2.9.1
billiard==3.6.4.0
blinker==1.4
boto3==1.17.112
botocore==1.20.112
cachetools==4.2.4
cattrs==1.5.0
celery==5.1.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
croniter==1.0.15
cryptography==36.0.0
decorator==5.1.0
defusedxml==0.7.1
dill==0.3.4
dnspython==2.1.0
docutils==0.16
email-validator==1.1.3
Flask==1.1.4
Flask-AppBuilder==3.4.0
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
flower==1.0.0
google-auth==2.3.3
graphviz==0.19
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.3
httpx==0.21.1
humanize==3.13.1
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
inflection==0.5.1
iso8601==1.0.2
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
jsonpath-ng==1.5.3
jsonschema==3.2.0
kombu==5.2.2
kubernetes==11.0.0
lazy-object-proxy==1.6.0
lockfile==0.12.2
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==1.1.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
numpy==1.21.4
oauthlib==3.1.1
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
packaging==21.3
pandas==1.3.3
pandas-redshift==2.0.5
pendulum==2.1.2
ply==3.11
prison==0.2.1
prometheus-client==0.12.0
prompt-toolkit==3.0.23
psutil==5.8.0
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.10.0
PyJWT==1.7.1
pyparsing==3.0.6
pyrsistent==0.18.0
python-daemon==2.3.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.3
pytzdata==2020.1
PyYAML==6.0
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rich==10.15.2
rsa==4.8
s3transfer==0.4.2
setproctitle==1.2.2
six==1.16.0
slack-sdk==3.12.0
sniffio==1.2.0
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
sqlalchemy-redshift==0.8.6
SQLAlchemy-Utils==0.37.8
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
tornado==6.1
unicodecsv==0.14.1
urllib3==1.26.7
vine==5.0.0
watchtower==1.0.6
wcwidth==0.2.5
websocket-client==1.2.3
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.6.0
